Fixes that type/feature mappings are not available during setup.
