+++ src/OrchardCore/OrchardCore.Abstractions/Extensions/IExtensionManager.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using OrchardCore.Environment.Extensions.Features;
@@ -9,7 +8,6 @@ namespace OrchardCore.Environment.Extensions
     {
         IExtensionInfo GetExtension(string extensionId);
         IEnumerable<IExtensionInfo> GetExtensions();
-        IEnumerable<Type> GetExportedExtensionTypes(IExtensionInfo extensionInfo);
         Task<ExtensionEntry> LoadExtensionAsync(IExtensionInfo extensionInfo);
 
         IEnumerable<IFeatureInfo> GetFeatures();
