+++ src/OrchardCore/OrchardCore/Shell/Builders/ShellContainerFactory.cs
@@ -152,49 +152,6 @@ namespace OrchardCore.Environment.Shell.Builders
 
         private void PopulateTypeFeatureProvider(ITypeFeatureProvider typeFeatureProvider, FeatureAwareServiceCollection featureAwareServiceCollection)
         {
-            // Get all types from all extension and add them to the type feature provider.
-            var extensions = _extensionManager.GetExtensions();
-
-            var allTypesByExtension = extensions
-                .SelectMany(extension =>
-                    _extensionManager.GetExportedExtensionTypes(extension)
-                        .Where(IsComponentType)
-                        .Select(type => new
-                        {
-                            Extension = extension,
-                            Type = type
-                        }));
-
-            var typesByFeature = allTypesByExtension
-                .GroupBy(typeByExtension => GetSourceFeatureNameForType(
-                    typeByExtension.Type,
-                    typeByExtension.Extension.Id))
-                .ToDictionary(
-                    group => group.Key,
-                    group => group.Select(typesByExtension => typesByExtension.Type));
-
-            foreach (var extension in extensions)
-            {
-                foreach (var feature in extension.Features)
-                {
-                    // Features can have no types.
-                    if (typesByFeature.TryGetValue(feature.Id, out var featureTypes))
-                    {
-                        // This is adding the types to the main feature for backward compatibility.
-                        // In the future we could stop doing it as we don't expect this to be necessary, and remove the FeatureTypeDiscovery attribute.
-                        foreach (var type in featureTypes)
-                        {
-                            // If the attribute is present then we explicitly ignore the backward compatibility and skip the registration
-                            // in the main feature.
-                            if (!SkipExtensionFeatureRegistration(type))
-                            {
-                                typeFeatureProvider.TryAdd(type, feature);
-                            }
-                        }
-                    }
-                }
-            }
-
             // Register all DIed types in ITypeFeatureProvider.
             foreach (var featureServiceCollection in featureAwareServiceCollection.FeatureCollections)
             {
@@ -222,23 +179,6 @@ namespace OrchardCore.Environment.Shell.Builders
                     }
                 }
             }
-        }
-
-        private static string GetSourceFeatureNameForType(Type type, string extensionId)
-        {
-            var attribute = type.GetCustomAttributes<FeatureAttribute>(false).FirstOrDefault();
-
-            return attribute?.FeatureName ?? extensionId;
-        }
-
-        private static bool IsComponentType(Type type)
-        {
-            return type.IsClass && !type.IsAbstract && type.IsPublic;
-        }
-
-        private static bool SkipExtensionFeatureRegistration(Type type)
-        {
-            return FeatureTypeDiscoveryAttribute.GetFeatureTypeDiscoveryForType(type)?.SkipExtension ?? false;
         }
     }
 }
