+++ test/OrchardCore.Tests/Extensions/ExtensionManagerTests.cs
@@ -31,6 +31,7 @@ namespace OrchardCore.Tests.Extensions
                 _applicationContext,
                 new[] { new ExtensionDependencyStrategy() },
                 new[] { new ExtensionPriorityStrategy() },
+                new TypeFeatureProvider(),
                 _moduleFeatureProvider,
                 new NullLogger<ExtensionManager>()
                 );
@@ -39,6 +40,7 @@ namespace OrchardCore.Tests.Extensions
                 _applicationContext,
                 new[] { new ExtensionDependencyStrategy() },
                 new[] { new ExtensionPriorityStrategy() },
+                new TypeFeatureProvider(),
                 _themeFeatureProvider,
                 new NullLogger<ExtensionManager>()
                 );
@@ -47,6 +49,7 @@ namespace OrchardCore.Tests.Extensions
                 _applicationContext,
                 new IExtensionDependencyStrategy[] { new ExtensionDependencyStrategy(), new ThemeExtensionDependencyStrategy() },
                 new[] { new ExtensionPriorityStrategy() },
+                new TypeFeatureProvider(),
                 _themeFeatureProvider,
                 new NullLogger<ExtensionManager>()
                 );
